Break up schema file into todo list/item apis
